+++ storage/overdata.c
@@ -174,7 +174,7 @@ overview_build(ARTNUM number, const char *article, size_t length,
     for (field = 0; field < ARRAY_SIZE(fields); field++) {
         buffer_append(overview, "\t", 1);
         if (field == 5) {
-            snprintf(buffer, sizeof(buffer), "%u", length);
+            snprintf(buffer, sizeof(buffer), "%lu", (unsigned long) length);
             buffer_append(overview, buffer, strlen(buffer));
         } else
             build_header(article, length, fields[field], overview);
